Fix OpenCL integer type widths in irif.h BUILTIN_CLASS functions
In OpenCL, 'long' is 64-bit and 'long long' is 128-bit (unlike C/C++).
The __builtin_generic_class_f32 used 'unsigned long' (64-bit) to alias
a float (32-bit), and __builtin_generic_class_f64 used 'unsigned long long'
(128-bit) to alias a double (64-bit). Both read beyond the value (UB) and
generated unnecessarily wide integer operations. The f64 variant created
i128 ops that crash the SPIR-V backend, and both bloated SPIR-V modules
causing translator timeouts.

Fix by using the correct-width types: 'unsigned int' (32-bit) for f32
and 'unsigned long' (64-bit) for f64.

+++ irif/inc/irif.h
@@ -170,7 +170,7 @@ static inline half __builtin_generic_clamp_f16(half x, half l, half h) {
 }
 
 static inline int __builtin_generic_class_f32(float x, int klass) {
-    unsigned long ix = *((unsigned long *)&x);
+    unsigned int ix = *((unsigned int *)&x);
 
     if ((klass & CLASS_PINF) && (ix == PINFBITPATT_SP32)) {
         return -1;
@@ -192,7 +192,7 @@ static inline int __builtin_generic_class_f32(float x, int klass) {
 }
 
 static inline int __builtin_generic_class_f64(double x, int klass) {
-    unsigned long long ix = *((unsigned long long *)&x);
+    unsigned long ix = *((unsigned long *)&x);
 
     if ((klass & CLASS_PINF) && (ix == PINFBITPATT_DP64)) {
         return -1;
